protovuf serializer test fix and cleanup
