+++ tests/test_protocols/test_base.py
@@ -16,22 +16,25 @@
 #   limitations under the License.
 #
 # ------------------------------------------------------------------------------
-
 """This module contains the tests of the messages module."""
 
 import os
 import shutil
 import tempfile
+from copy import copy
 from enum import Enum
 from pathlib import Path
 from typing import Callable, List, Tuple, Type
 
 import pytest
+from google.protobuf.struct_pb2 import Struct
 
 from aea.configurations.constants import DEFAULT_PROTOCOL
 from aea.exceptions import AEAEnforceError
 from aea.mail.base import Envelope
-from aea.protocols.base import Message, ProtobufSerializer, Protocol
+from aea.mail.base_pb2 import DialogueMessage as Pb2DialogueMessage
+from aea.mail.base_pb2 import Message as ProtobufMessage
+from aea.protocols.base import Message, Protocol, Serializer
 from aea.protocols.dialogue.base import Dialogue, DialogueLabel
 
 from packages.fetchai.protocols.default.dialogues import (
@@ -138,6 +141,90 @@ class TestMessageProperties:
         )
 
 
+class ExampleProtobufSerializer(Serializer):
+    """
+    Example Protobuf serializer.
+
+    It assumes that the Message contains a JSON-serializable body.
+    """
+
+    @staticmethod
+    def encode(msg: Message) -> bytes:
+        """
+        Encode a message into bytes using Protobuf.
+
+        - if one of message_id, target and dialogue_reference are not defined,
+          serialize only the message body/
+        - otherwise, extract those fields from the body and instantiate
+          a Message struct.
+        """
+        message_pb = ProtobufMessage()
+        if msg.has_dialogue_info:
+            dialogue_message_pb = Pb2DialogueMessage()
+            dialogue_message_pb.message_id = msg.message_id
+            dialogue_message_pb.dialogue_starter_reference = msg.dialogue_reference[0]
+            dialogue_message_pb.dialogue_responder_reference = msg.dialogue_reference[1]
+            dialogue_message_pb.target = msg.target
+
+            new_body = copy(msg._body)  # pylint: disable=protected-access
+            new_body.pop("message_id")
+            new_body.pop("dialogue_reference")
+            new_body.pop("target")
+
+            body_json = Struct()
+            body_json.update(new_body)  # pylint: disable=no-member
+
+            dialogue_message_pb.content = (  # pylint: disable=no-member
+                body_json.SerializeToString()
+            )
+            message_pb.dialogue_message.CopyFrom(  # pylint: disable=no-member
+                dialogue_message_pb
+            )
+        else:
+            body_json = Struct()
+            body_json.update(msg._body)  # pylint: disable=no-member,protected-access
+            message_pb.body.CopyFrom(body_json)  # pylint: disable=no-member
+
+        return message_pb.SerializeToString()
+
+    @staticmethod
+    def decode(obj: bytes) -> Message:
+        """
+        Decode bytes into a message using Protobuf.
+
+        First, try to parse the input as a Protobuf 'Message';
+        if it fails, parse the bytes as struct.
+        """
+        message_pb = ProtobufMessage()
+        message_pb.ParseFromString(obj)
+        message_type = message_pb.WhichOneof("message")
+        if message_type == "body":
+            body = dict(message_pb.body)  # pylint: disable=no-member
+            msg = TestMessage(_body=body)
+            return msg
+        if message_type == "dialogue_message":
+            dialogue_message_pb = (
+                message_pb.dialogue_message  # pylint: disable=no-member
+            )
+            message_id = dialogue_message_pb.message_id
+            target = dialogue_message_pb.target
+            dialogue_starter_reference = dialogue_message_pb.dialogue_starter_reference
+            dialogue_responder_reference = (
+                dialogue_message_pb.dialogue_responder_reference
+            )
+            body_json = Struct()
+            body_json.ParseFromString(dialogue_message_pb.content)
+            body = dict(body_json)
+            body["message_id"] = message_id
+            body["target"] = target
+            body["dialogue_reference"] = (
+                dialogue_starter_reference,
+                dialogue_responder_reference,
+            )
+            return TestMessage(_body=body)
+        raise ValueError("Message type not recognized.")  # pragma: nocover
+
+
 class TestBaseSerializations:
     """Test that the base serializations work."""
 
@@ -155,7 +242,7 @@ class TestBaseSerializations:
 
     def test_default_protobuf_serialization(self):
         """Test that the default Protobuf serialization works."""
-        message_bytes = ProtobufSerializer().encode(self.message)
+        message_bytes = ExampleProtobufSerializer().encode(self.message)
         envelope = Envelope(
             to="receiver",
             sender="sender",
@@ -168,13 +255,13 @@ class TestBaseSerializations:
         actual_envelope = envelope
         assert expected_envelope == actual_envelope
 
-        expected_msg = ProtobufSerializer().decode(expected_envelope.message)
+        expected_msg = ExampleProtobufSerializer().decode(expected_envelope.message)
         actual_msg = self.message
         assert expected_msg == actual_msg
 
     def test_default_protobuf_serialization_with_dialogue_info(self):
         """Test that the default Protobuf serialization with dialogue info works."""
-        message_bytes = ProtobufSerializer().encode(self.message3)
+        message_bytes = ExampleProtobufSerializer().encode(self.message3)
         envelope = Envelope(
             to="receiver",
             sender="sender",
@@ -187,7 +274,7 @@ class TestBaseSerializations:
         actual_envelope = envelope
         assert expected_envelope == actual_envelope
 
-        expected_msg = ProtobufSerializer().decode(expected_envelope.message)
+        expected_msg = ExampleProtobufSerializer().decode(expected_envelope.message)
         actual_msg = self.message3
         assert expected_msg == actual_msg
